Refactor class back to cryptotransfer
